Модуляция
В модуль util.js поместила: вспомогательные функции.
В модуле data.js поместила код, который отвечает за создание данных
В модуле main.js импортировала функции из других модулей и оставила код, который является точкой входа
